requirements.txt
# Auteur : Gilles - Projet : AGC Space - Module : Dépendances
Django==5.2.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
django-redis==6.0.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
Pillow==12.2.0
gunicorn==25.3.0
redis==6.2.0
PyJWT==2.10.1

# Media processing
moviepy==1.0.3
python-magic==0.4.27

# Tests
pytest==9.0.3
pytest-django==4.12.0
pytest-mock==3.15.1
pytest-cov==6.2.1
